requirements.txt
python-telegram-bot==20.3
yt-dlp==2023.10.07
aiofiles==23.1.0
python-dotenv==1.0.0
redis==5.0.9
